Move arena code to arena.go.
